@@ -4,7 +4,7 @@
 	"description": "Timings for HTTP requests",
 	"main": "dist",
 	"engines": {
-		"node": ">=8"
+		"node": ">=10"
 	},
 	"scripts": {
 		"test": "xo && nyc ava",
@@ -32,7 +32,7 @@
 	},
 	"homepage": "https://github.com/szmarczak/http-timer#readme",
 	"dependencies": {
-		"defer-to-connect": "^1.0.2"
+		"defer-to-connect": "^1.1.0"
 	},
 	"devDependencies": {
 		"@sindresorhus/tsconfig": "^0.3.0",
